+++ paifeaturestore-20230621/alibabacloud_paifeaturestore20230621/models/_create_label_table_request.py
@@ -73,10 +73,12 @@ class CreateLabelTableRequest(DaraModel):
 class CreateLabelTableRequestFields(DaraModel):
     def __init__(
         self,
+        aligned_entity_name: str = None,
         attributes: List[str] = None,
         name: str = None,
         type: str = None,
     ):
+        self.aligned_entity_name = aligned_entity_name
         # This parameter is required.
         self.attributes = attributes
         # This parameter is required.
@@ -92,6 +94,9 @@ class CreateLabelTableRequestFields(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.aligned_entity_name is not None:
+            result['AlignedEntityName'] = self.aligned_entity_name
+
         if self.attributes is not None:
             result['Attributes'] = self.attributes
 
@@ -105,6 +110,9 @@ class CreateLabelTableRequestFields(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AlignedEntityName') is not None:
+            self.aligned_entity_name = m.get('AlignedEntityName')
+
         if m.get('Attributes') is not None:
             self.attributes = m.get('Attributes')