+++ paifeaturestore-20230621/alibabacloud_paifeaturestore20230621/models/_get_label_table_response_body.py
@@ -125,10 +125,12 @@ class GetLabelTableResponseBody(DaraModel):
 class GetLabelTableResponseBodyFields(DaraModel):
     def __init__(
         self,
+        aligned_entity_name: str = None,
         attributes: List[str] = None,
         name: str = None,
         type: str = None,
     ):
+        self.aligned_entity_name = aligned_entity_name
         self.attributes = attributes
         self.name = name
         self.type = type
@@ -141,6 +143,9 @@ class GetLabelTableResponseBodyFields(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.aligned_entity_name is not None:
+            result['AlignedEntityName'] = self.aligned_entity_name
+
         if self.attributes is not None:
             result['Attributes'] = self.attributes
 
@@ -154,6 +159,9 @@ class GetLabelTableResponseBodyFields(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AlignedEntityName') is not None:
+            self.aligned_entity_name = m.get('AlignedEntityName')
+
         if m.get('Attributes') is not None:
             self.attributes = m.get('Attributes')